Updated Authentication, login and error
This app now has login capabilities. I had issues with users being able to login in without an account and still get to a page but no data or info was found/ populated because it didnt exist in the database

Moving Forward the app needs to update the UI altogether, edit the check-in view to populate users and the the users set bus stops with GPS to the locations.